component enumerator working
