+++ biocrnpyler/__init__.py
@@ -36,3 +36,4 @@ from .compartments import *
 from .utils import *
 
 #checking for nonexistant plotting-related modules now happens in plotting.py
+from .component_enumerator import *
